+++ Assembly-CSharp/Engine/Source/Blenders/LerpBlendOperation.cs
@@ -1,4 +1,5 @@
 ﻿using Engine.Common.Blenders;
+using UnityEngine;
 
 namespace Engine.Source.Blenders
 {
@@ -10,7 +11,7 @@ namespace Engine.Source.Blenders
 
     public Vector2 Blend(Vector2 a, Vector2 b) => Vector2.Lerp(a, b, Time);
 
-    public int Blend(int a, int b) => (int) Mathf.Lerp((float) a, (float) b, Time);
+    public int Blend(int a, int b) => (int) Mathf.Lerp(a, b, Time);
 
     public float Blend(float a, float b) => Mathf.Lerp(a, b, Time);
   }
