+++ Assembly-CSharp/Expressions/PowIntOperation.cs
@@ -1,5 +1,6 @@
 ﻿using Engine.Common.Generator;
 using Inspectors;
+using UnityEngine;
 
 namespace Expressions
 {
@@ -7,7 +8,7 @@ namespace Expressions
   [GenerateProxy(TypeEnum.Cloneable | TypeEnum.Copyable | TypeEnum.DataRead | TypeEnum.DataWrite)]
   public class PowIntOperation : BinaryOperation<int>
   {
-    protected override int Compute(int a, int b) => (int) Mathf.Pow((float) a, (float) b);
+    protected override int Compute(int a, int b) => (int) Mathf.Pow(a, b);
 
     protected override string OperatorView() => "^";
   }
